layout: somewhat less broken image sizing
This makes images a property of BlockBox instead of a special case of
InlineBox, which makes it way more convenient to implement CSS sizing.
(Representing img as BlockBox is made feasible by the fact that replaced
inline-block is sized the same way as replaced inline.)

Also, the sizing algorithm now complies with CSS 2.1.  At least I *think*
it does; it's hard to tell because the spec uses a halfway ambiguous and
fairly counter-intuitive representation of what appear to be some rather
simple rules.

Flex image sizing still seems broken in some places, but this might have
more to do with deficiencies in our flex implementation than images...